feat: add option to control the pending channel join min rate
The default value is set to the join rate limit described in Twitch documentation "20 join attempts per 10 seconds per user" or every 500ms for normal accounts (or anonymous). I know that joins can be sent faster than every 500ms, but this will ensure a minimum rate that won't get dropped for rate reasons. Ping time is not taken into account, so maybe some jitter or similar should be added. There is a concurrency limit of 100 channels but I've also observed it to not be a hard limit. Accounts can be verified to have better rates and higher limits. See https://dev.twitch.tv/docs/chat/#join-rate-limits